application/pyproject.toml
[project]
name = "application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "blinker>=1.9.0",
    "chromadb>=1.3.5",
    "docling-core[chunking]>=2.52.0",
    "environ-config>=24.1.0",
    "inquirer>=3.4.1",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-docling>=2.0.0",
    "langchain-experimental>=0.4.0",
    "langchain-milvus>=0.3.1",
    "langchain-mongodb>=0.8.0",
    "ollama-ocr>=0.1.6",
    "openparse>=0.7.0",
    "panda>=0.3.1",
    "pdfminer-six>=20251107",
    "pymilvus>=2.6.4",
    "pymongo>=4.15.4",
    "pymupdf>=1.26.6",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "streamlit-authenticator>=0.4.2",
    "streamlit-pdf-viewer>=0.0.26",
    "unstructured[pdf]>=0.18.21",
    "uuid>=1.30",
]
